server/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-rest-knox=4.2.0
pytest==7.4.0
pytest-django==4.5.2
